Arquetipo Java

+++ order-commons/src/main/java/mx/com/axity/commons/to/KeyboardTO.java
@@ -10,7 +10,7 @@ import java.io.Serializable;
 @Setter
 public class KeyboardTO implements Serializable {
 
-    private Long idKeyboard;
+    private Integer idKeyboard;
     private String brand;
     private String model;
     private String serialNumber;
